// src/openapi/model/changerMotDePasseUtilisateurDto.ts
/**
 * Gestion de stock REST API
 * Documentation pour l\'API REST Gestion des stocks
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: wiamejaoui.wj@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChangerMotDePasseUtilisateurDto { 
    id?: number;
    motDePasse?: string;
    confirmMotDePasse?: string;
}